Upgrade workspace dependencies to latest (except bincode, pgwire)
Bump all dependencies to their latest releases and pin Cargo.toml to the
resolved versions, holding back two for safety: bincode (1->3 would change
the on-disk SSTable format) and pgwire (0.23->0.40 is a large wire-layer
rewrite). Notable majors taken: sqlparser 0.45->0.62, axum 0.7->0.8,
reqwest 0.12->0.13, prometheus-client 0.22->0.25, opentelemetry 0.27->0.32,
tower-http 0.5->0.7, thiserror 1->2, sha2 0.11, object_store 0.14,
rcgen 0.14, tokio-postgres-rustls 0.14.

API migrations:
- sqlparser 0.62: Statement variants now wrap structs (CreateTable, Insert,
  AlterTable, Set/SetTimeZone); JSON access is BinaryOp with arrow operators;
  Function args are FunctionArguments; ORDER BY/LIMIT and Query fields moved;
  CASE uses CaseWhen; JoinOperator gained Join/Left/Right plus tuple CrossJoin;
  SUBSTR parses to a dedicated Substring node. INSERT column names are taken
  from the bare identifier (not to_string(), which re-quotes "Id") so quoted
  client SQL (EF Core) matches the catalog's unquoted columns.
- axum 0.8: route params are {name}, not :name.
- reqwest 0.13: rustls-tls/query are split features; tls_certs_only replaces
  tls_built_in_root_certs(false)+add_root_certificate.
- object_store 0.14: put/get/head/delete moved to ObjectStoreExt; size is u64.
- sha2 0.11: digest output hex-encoded manually.
- opentelemetry 0.32: SdkTracerProvider; batch exporter drops the runtime arg.
- rcgen 0.14: signed_by takes an Issuer; CertifiedKey.signing_key.

Tests run fast Raft timers (testkit) as before. Verified: full --all-targets
build, clippy -D warnings, and a broad test cross-section (executor, sql,
import, telemetry, backup, security, and pg18/pgwire/jdbc/npgsql compat).

+++ crates/nodus_backup/src/s3.rs
@@ -6,7 +6,7 @@ use async_trait::async_trait;
 use bytes::Bytes;
 use chrono::Utc;
 use futures_util::StreamExt;
-use object_store::{ObjectStore, aws::AmazonS3Builder, path::Path as ObjPath};
+use object_store::{ObjectStore, ObjectStoreExt, aws::AmazonS3Builder, path::Path as ObjPath};
 use std::sync::Arc;
 
 /// Connection settings for an S3-compatible object store.
@@ -101,7 +101,7 @@ impl BackupRepository for S3BackupRepository {
             let meta = meta?;
             out.push(ObjectMetadata {
                 key: meta.location.to_string(),
-                size: meta.size as u64,
+                size: meta.size,
                 last_modified: meta.last_modified,
             });
         }
